test: move HttpResponseMessage helper to the right namespace
