--- Doppler.HtmlEditorApi.Test/HttpResponseMessageExtensions.cs
@@ -1,15 +1,0 @@
-using System.Linq;
-using System.Net.Http;
-
-namespace Doppler.HtmlEditorApi
-{
-    public static class HttpResponseMessageExtensions
-    {
-        public static string GetHeadersAsString(this HttpResponseMessage response)
-        {
-            var keysAndValues = response.Headers.SelectMany(x => x.Value.Select(y => new { x.Key, Value = y }));
-            var headerLines = keysAndValues.Select(x => $"{x.Key}: {x.Value}");
-            return string.Join("\n", headerLines);
-        }
-    }
-}
