+++ Doppler.HtmlEditorApi.Test/Utils/HttpResponseMessageExtensions.cs
@@ -1,0 +1,14 @@
+using System.Linq;
+using System.Net.Http;
+
+namespace Doppler.HtmlEditorApi.Test.Utils;
+
+public static class HttpResponseMessageExtensions
+{
+    public static string GetHeadersAsString(this HttpResponseMessage response)
+    {
+        var keysAndValues = response.Headers.SelectMany(x => x.Value.Select(y => new { x.Key, Value = y }));
+        var headerLines = keysAndValues.Select(x => $"{x.Key}: {x.Value}");
+        return string.Join("\n", headerLines);
+    }
+}
